Move player to it's own component.
Fixes a Vue specific bug, as well as captions.
